@@ -14,7 +14,9 @@
     "start": "webpack-dev-server --config ./configs/webpack/dev.config.js",
     "test": "jest --ci",
     "test:local": "jest",
-    "cover": "jest --coverage"
+    "cover": "jest --coverage",
+    "lint": "set -ex; eslint --ext .js,.ts,.tsx .",
+    "lint:fix": "npm run lint -- --fix"
   },
   "dependencies": {
     "@solana/web3.js": "^0.29.4",
@@ -60,12 +62,16 @@
     "@types/react-select": "^3.0.5",
     "@types/react-share": "^3.0.3",
     "@types/react-test-renderer": "^16.0.3",
+    "@typescript-eslint/eslint-plugin": "^2.17.0",
+    "@typescript-eslint/parser": "^2.17.0",
     "args": "^5.0.1",
     "babel-loader": "^8.0.6",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^3.2.0",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.11.2",
+    "eslint": "^6.8.0",
+    "eslint-plugin-react": "^7.18.0",
     "file-loader": "^4.1.0",
     "fork-ts-checker-webpack-plugin": "^1.5.0",
     "html-webpack-plugin": "^4.0.0-beta.8",
@@ -88,7 +94,7 @@
     "ts-jest": "^25.0.0",
     "ts-loader": "^6.0.4",
     "ts-mockito": "^2.3.1",
-    "typescript": "^3.5.3",
+    "typescript": "^3.7.5",
     "url-loader": "^2.1.0",
     "webpack": "^4.39.1",
     "webpack-dev-server": "^3.7.2"
