@@ -1,6 +1,6 @@
-import {RESET_STATISTICS_GAME} from "../types";
+import {Action, RESET_STATISTICS_GAME} from "../types";
 
-export const resetStatisticsGame = () => {
+export const resetStatisticsGame = (): Action => {
     return {
         type: RESET_STATISTICS_GAME,
     }
