@@ -1,7 +1,7 @@
-import {SET_STATISTICS_GAME} from "../types";
+import {Action, SET_STATISTICS_GAME} from "../types";
 import IGame from "../../reducers/game/model";
 
-export const setStatisticsGame = (statistics: IGame.ModelStatistics) => {
+export const setStatisticsGame = (statistics: IGame.ModelStatistics): Action => {
     return {
         type: SET_STATISTICS_GAME,
         payload: statistics
