@@ -1,7 +1,7 @@
-import {SET_INFO} from "../types";
+import {Action, SET_INFO} from "../types";
 import ITransaction from "../../reducers/transactions/model";
 
-export const setTransactionInfo = (info: ITransaction.Model) => {
+export const setTransactionInfo = (info: ITransaction.Model): Action => {
     return {
         type: SET_INFO,
         payload: info
