@@ -1,6 +1,6 @@
-import {SET_TPS} from "../types";
+import {Action, SET_TPS} from "../types";
 
-export const setTransactionsPerSecond = (tps: number) => {
+export const setTransactionsPerSecond = (tps: number): Action => {
     return {
         type: SET_TPS,
         payload: tps,
