@@ -3,8 +3,8 @@ import * as React from 'react';
 import './index.scss';
 
 interface IProps {
-    dayTransactionCounts: number
-    gameTransactionCounts: number
+    dayTransactionCounts: number;
+    gameTransactionCounts: number;
 }
 
 export const GlobalStatisticsBoard = ({dayTransactionCounts, gameTransactionCounts}: IProps) => {
