@@ -1,23 +1,18 @@
-
-namespace ITransaction {
-    export interface Model {
-        id: string
-        status: string
-        info: TransactionInfo
-    }
-
-    export interface TransactionInfo {
-        signature: string
-        confirmationTime: number
-        lamportsCount: number
-    }
-
-    export interface ModelState {
-        transactions: Model[]
-        countCompletedTransactions: number
-        averageTransactionsTime: number
-        transactionsPerSecond: number
-    }
+export interface Model {
+    id: string;
+    status: string;
+    info: TransactionInfo;
 }
 
-export default ITransaction;
+export interface TransactionInfo {
+    signature: string;
+    confirmationTime: number;
+    lamportsCount: number;
+}
+
+export interface ModelState {
+    transactions: Model[];
+    countCompletedTransactions: number;
+    averageTransactionsTime: number;
+    transactionsPerSecond: number;
+}
