@@ -1,8 +1,8 @@
 export interface IGameService {
-    saveGame(data: any): Promise<any>
-    getDailyGameCounts(): any
-    getDailyTransactionCounts(): any
-    getGameTransactionCounts(): any
+    saveGame(data: any): Promise<any>;
+    getDailyGameCounts(): any;
+    getDailyTransactionCounts(): any;
+    getGameTransactionCounts(): any;
 }
 
 
