@@ -3,7 +3,7 @@ import {IGameService} from "./game-service/model";
 import {IDefaultWebSocketService} from "./web-socket/model";
 
 export interface IService {
-    transactionsService: ITransactionsService
-    gameService: IGameService
-    wsService: IDefaultWebSocketService
+    transactionsService: ITransactionsService;
+    gameService: IGameService;
+    wsService: IDefaultWebSocketService;
 }
