@@ -1,16 +1,16 @@
-const required = (value: string) => {
+const required = (value: string): boolean => {
     return (value.length > 0);
 };
 
-const email = (value: string) => {
-    return value.toLowerCase().match(/^([\w-]+\.)*[\w-]+@[a-z0-9_-]+(\.[a-z0-9_-]+)*\.[a-z]{2,6}$/)
+const email = (value: string): boolean => {
+    return value.toLowerCase().match(/^([\w-]+\.)*[\w-]+@[a-z0-9_-]+(\.[a-z0-9_-]+)*\.[a-z]{2,6}$/) !== null;
 };
 
-const minLength = (value: string, min: number) => {
+const minLength = (value: string, min: number): boolean => {
     return (value.length >= min)
 };
 
-const valueIsEmpty = (value: string) => {
+const valueIsEmpty = (value: string): boolean => {
     return value.trim().length === 0
 };
 
